+++ src/qibo/tensorflow/cgates.py
@@ -617,7 +617,7 @@ class Unitary(MatrixGate, gates.Unitary):
         unitary = self.parameters
         if isinstance(unitary, qnp.Tensor):
             return qnp.cast(unitary)
-        if isinstance(unitary, self.K.Tensor):
+        if isinstance(unitary, K.Tensor):
             return K.copy(K.cast(unitary))
 
     def _dagger(self) -> "Unitary":
